config: version 4.6-Sate
Satay (sate in Bahasa Indonesia1) is a dish similar to kebabs in that it is made of cubes of skewered meat that is grilled and eaten with a peanut sauce dip.2
Tracing its origins to the Arabs,3 satay has been adapted to the multicultural palates of Asians, with various spicy sauces and different ways of marinating the meat.4

'Origins'
Arabs were known to skewer their meat on swords before roasting it and Middle Eastern nomads would barbecue their meat on metal skewers in a dish known as kebab.5 The spice trade brought Arab traders to Southeast Asia, which led to the spread of Arabic cuisine to Indonesia, and eventually to Malaya.6 The kebab can also be found in India and other countries such as Greece and Turkey.7
However, a key adaptation of the dish in Asia is that wooden rather than metal skewers are used.8

Satay sauce is made from ground peanuts, and other spices such as coriander and cumin seeds.9
Various seasonings are used to marinate the pieces of meat and the specially-made peanut sauce is usually served as a dip together with the grilled meat.10

'Description'
The commonly used meats in satay are beef, mutton, lamb and chicken.11 Among non-Muslims, pork is also used.12 The small cuts of meat are marinated in various spices that also tenderise the meat.13 They are then skewered through sharp, wooden sticks.14 Dried, thin stems of the coconut leaf were originally used as satay sticks.15 But, these days bamboo sticks are used.16
The satay is barbecued over a flaming charcoal fire and constantly brushed with oil for a tantalising glaze, until well-browned.17 The sticks of grilled meat are then served with a bowl of peanut dip and cuts of cucumber and onion.18 Several sticks can be eaten at one go due to the small, bite-sized pieces of meat. Satay is often accompanied by ketupat (steamed rice) wrapped in woven coconut leaf packets.19
Some types of satay come with a much sweeter flavour with a twirl of kechap manis (sweet sauce) added to the peanut sauce.20 Other spices such as galangal (a type of ginger) and finely cut dashes of the limau purut (kaffir lime) leaf further heighten the flavour of the satay and its sauce.21

ref. https://eresources.nlb.gov.sg/infopedia/articles/SIP_888_2005-01-10.html